Create context for Alerts for readability and optimization
